@@ -57,7 +57,7 @@
     "coveralls": " ~3.0.0",
     "cwd": "~0.10.0",
     "doctoc": " ~1.3.0",
-    "eslint": "~4.7.2",
+    "eslint": "~4.8.0",
     "eslint-config-airbnb-base": "~12.0.0",
     "eslint-plugin-filenames": "1.2.0",
     "eslint-plugin-import": " ~2.7.0",
@@ -72,7 +72,7 @@
     "jsdoc-to-markdown": "~3.0.0",
     "nsp": "~2.8.0",
     "package-json-to-readme": "~2.0.0",
-    "winston": "~2.3.1"
+    "winston": "^2.4.0"
   },
   "preferGlobal": true,
   "bin": {
