@@ -26,25 +26,25 @@ describe(TEST_SUITE_DESCRIPTION_UNIT + ' - serialize-utils - ', () => {
   describe('Function createExportString', () => {
     it('should create ES6 default export', async () => {
       expect.assertions(1);
-      const result = await createExportString(true);
+      const result = await createExportString(false);
       expect(result).toBe('export default ');
     });
 
     it('should create "module.exports"', async () => {
       expect.assertions(1);
-      const result = await createExportString(false);
+      const result = await createExportString(true);
       expect(result).toBe('module.exports = ');
     });
 
     it('should create ES6 default export with named export', async () => {
       expect.assertions(1);
-      const result = await createExportString(true, 'foo');
+      const result = await createExportString(false, 'foo');
       expect(result).toBe(`export const ${namedExport} = `);
     });
 
     it('should create "module.exports" with named export', async () => {
       expect.assertions(1);
-      const result = await createExportString(false, 'foo');
+      const result = await createExportString(true, 'foo');
       expect(result).toBe(`module.exports.${namedExport} = `);
     });
   });
@@ -55,7 +55,7 @@ describe(TEST_SUITE_DESCRIPTION_UNIT + ' - serialize-utils - ', () => {
       const result = await serializeJsToString(toSerializeToJs, {
         strict: true,
         indent: indent.length,
-        'no-es6': false,
+        es5: false,
       });
       expect(result).toBe(
         `'use strict;'${nl}${nl}export default {${nl}${indent}foo: 'bar',${nl}${indent}bar: {bar: 'bar'}${nl}};${nl}`
@@ -67,7 +67,7 @@ describe(TEST_SUITE_DESCRIPTION_UNIT + ' - serialize-utils - ', () => {
       const result = await serializeJsToString(toSerializeToJs, {
         strict: false,
         indent: indent.length,
-        'no-es6': false,
+        // es5: false,
       });
       expect(result).toBe(`export default {${nl}${indent}foo: 'bar',${nl}${indent}bar: {bar: 'bar'}${nl}};${nl}`);
     });
@@ -77,11 +77,11 @@ describe(TEST_SUITE_DESCRIPTION_UNIT + ' - serialize-utils - ', () => {
       const result = await serializeJsToString(toSerializeToJs, {
         strict: true,
         indent: indent.length,
-        'no-es6': false,
-        'no-single': true,
+        es5: true,
+        double: true,
       });
       expect(result).toBe(
-        `"use strict;"${nl}${nl}export default {${nl}${indent}foo: "bar",${nl}${indent}bar: {bar: "bar"}${nl}};${nl}`
+        `"use strict;"${nl}${nl}module.exports = {${nl}${indent}foo: "bar",${nl}${indent}bar: {bar: "bar"}${nl}};${nl}`
       );
     });
   });
